completed

+++ app/Http/Controllers/CategoryController.php
@@ -96,6 +96,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0){
+            session()->flash('error', 'Category cannot deleted because it has some posts');
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash('success', 'Category deleted successfully');
         return redirect()->back();
