+++ app/Http/Controllers/TagController.php
@@ -100,6 +100,11 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('error', 'Tag cannot be deleted because it associated with some post');
+            return redirect()->back();
+        }
+
         $tag->delete();
         session()->flash('success', 'Tag deleted successfully');
         return redirect()->back();
